Configure some linter settings for TYPO3 and add PhpUnit bootstrap
Summary:
Changed some linter settings to ignore a linter
that checks for the file ending "php" because files are
piped over STDIN which will make this lint fail.
Also added php files to the ingore list of the "text"
linter because these checks do not match the TYPO3 CGL
linters.

Added also the missing phpunit.xml and bootstrap.php file.

Reviewers: #forgetit-dev, claus.due

Reviewed By: #forgetit-dev, claus.due

Subscribers: claus.due

Differential Revision: https://phabricator.dkd.de/D102